pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "spl-lang"
version = "0.1.0"
description = "Structured Prompt Language - SQL for LLM Context Management"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Wen Gong"},
]
keywords = ["llm", "prompt-engineering", "sql", "declarative", "context-management"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "faiss-cpu>=1.7.4",
    "numpy>=1.24",
    "httpx>=0.25",
    "tiktoken>=0.5",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
]

[project.scripts]
spl = "spl.cli:main"

[project.urls]
Homepage = "https://github.com/digital-duck/SPL"
Repository = "https://github.com/digital-duck/SPL"

[tool.setuptools.packages.find]
include = ["spl*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
